+++ packages/person/src/index.tsx
@@ -1,6 +1,7 @@
 export * from './PersonAvatar';
 export * from './PersonCard';
 export * from './PersonProvider';
+export * from './PersonListItem';
 export type { PersonResolver } from '@equinor/fusion-wc-person/person-provider';
 export { PersonAvailability, PersonAccountType } from '@equinor/fusion-wc-person';
 export type { PersonDetails, PersonPresence, PersonPicture } from '@equinor/fusion-wc-person';